changes to CFEB03

+++ emuDQM/EmuTests/src/common/Test_CFEB03.cc
@@ -726,7 +726,7 @@ void Test_CFEB03::finishCSC(std::string cscID)
 	    res_out << std::fixed << std::setprecision(2) <<  (first_strip_index+layer*strips_per_layer+strip) << "  "
 		    << r01.content[layer][strip]  << "  " << r02.content[layer][strip] << "  " <<  r03.content[layer][strip] 
 		    << std::setprecision(3) << "  " << r04.content[layer][strip] << "  " << r05.content[layer][strip] 
-			<< "  " << mask.content[layer][strip] << std::endl;
+			<< "  " << (int)(mask.content[layer][strip]) << std::endl;
 	  }
 	}
 	res_out.close();
@@ -738,7 +738,7 @@ void Test_CFEB03::finishCSC(std::string cscID)
             res_out << std::fixed << std::setprecision(5) <<  (first_strip_index+layer*strips_per_layer+strip) << "  "
                     << r06.content[layer][strip]  << "  " << r07.content[layer][strip] << "  "  
                     << r09.content[layer][strip] << "  " << r10.content[layer][strip] 
-			<< "  " << mask.content[layer][strip] << std::endl;
+			<< "  " << (int)(mask.content[layer][strip]) << std::endl;
           }
         }
         res_out.close();
